🔧 Replace Flask config handler with an simpler Dataclass (#447)
* 💥 Replace config handler.
Move out of the Flask config in favour for a simple Dataclass.
The new config handler should be easier to use around the app, but it does not support overrides as the old did.
As a result of no overrides I've had to remove the one test that relied on the override functionality, not a huge loss.
Also config from file is now out, only env

* ♻️ Use new configuration handler

* 🗑️ Remove now unsupported config override from tests

* 🔥 Remove unsupported test

* 🚨 Fix linter warnings

* 👷 Add missing config in GitHub Actions for unittests